# [Core] Optimize the computation of prefix hashes for chunks